fix: scopes args file
This PR changes the proto plugin to take into account the name of target
being built to generate the args file. Before this patch, the rules that
use the plugin will generate clashing args files (`path/to/rule/args`).
This is fine if you run only one test at a time, but if you happen to
have a `_lint` and a `_breaking` test in the same bazel package, an
error pops up because both targets are trying to execute actions writing
the same file.