Fix running tests in Windows (#79)
This should resolve the toolchain registration issue from #78.

Also, it adds a batch script version of the shell script shim used to
invoke protoc for tests; however, this requires `--enable_runfiles` to
be explicitly passed or added to `.bazelrc` for a project, as Bazel on
Windows does not populate runfiles by default.

In the future, we will need a better approach to supporting Windows that
uses the runfiles manifest instead. This is tricky though, and will
require some careful thought.