Migrate to new BCR release workflow (#116)
This follows the work that I did for protovalidate and protovalidate-cc
to do the same thing, albeit it is a little simpler here. I've tested
the adapted release_prep.sh and it LGTM. Still, we might need a couple
releases to actually figure it out and get attestation working
correctly.

Note: before this can be deployed, we need to also clean up the old BCR
app from this repository.