Fix: pass in the repo root for the 'dir' argument (#119)
In tools that run bazel commands from within `bazel run` the
`BUILD_WORKSPACE_DIRECTORY` will be set and then in this
[code](https://github.com/bazel-contrib/bazel-gazelle/blob/master/cmd/gazelle/fix-update.go#L122-L139)
the directory will eventually be set to the top level repo root. An
example of a tool that this can happen to is the
[gopackagesdriver](https://github.com/bazel-contrib/rules_go/wiki/Editor-setup)
in rules_go.

This results in error messages like:
```
 no such package '@@rules_buf++buf+buf_deps//buf/validate': failed with code: 1, error: gazelle: .: not a subdirectory of repo root
```

By passing in additional argument it explicitly sets the `dir` to the
directory of the repo. Gazelle itself does something
[similar](https://github.com/bazel-contrib/bazel-gazelle/blob/6803436a9c39209bd56ea1dba2dddda323a3c98b/internal/go_repository.bzl#L363)

Co-authored-by: jchadwick-buf <116005195+jchadwick-buf@users.noreply.github.com>